Docker-Swarm/ansible/requirements.txt
azure-common
azure-mgmt-compute
